+++ src/h.js
@@ -699,6 +699,17 @@ var H = (function() {
     return parseInt(v, 10);
   };
 
+  this.getj = function(start, sel/*, default */) {
+
+    var d = undefined;
+    var a = self.toArray(arguments);
+    var l = a[a.length - 1];
+    if (a.length > 2) { d = a[2]; }
+    else if (a.length > 1 && (typeof l) !== 'string') { d = a.pop(); }
+    var v = self.get.apply(null, a);
+    try { return JSON.parse(v); } catch(e) { return d; }
+  };
+
   this.set = function(start, sel, value) {
 
     var a = self.toArray(arguments);